PortfolioLayoutからMeta関連を分離
